app/requirements.txt
python-dotenv
fastapi
uvicorn[standard]
pydantic
openai
redis